Create doubleexprg.kt
Three variables called trip1, trip2, and trip3 are declared to represent the distance of each part of the trip. All of them are Double variables because they store decimal values. Use val to declare each variable because their values do not change over the course of the program. The program also creates a fourth variable called totalTripLength which is currently initialized to 0.0. The last line of the program prints a message with the value of the totalTripLength.