Create prg.kt
Notice that there are two String variables declared, a nextMeeting variable and a date variable. Then a third String variable called reminder is declared, which is set equal to the nextMeeting variable plus the date variable.

With the + symbol, you can add two strings together, which is called concatenation. The two strings are combined together, one after the other. The result of the expression, nextMeeting + date is "Next meeting is:January 1".
The value "Next meeting is:January 1" is then stored into the reminder variable using the assignment operator on line 4 of the program.